refactor(clinic): rename clinic_services to clinix_services
